requirements.txt
flask
requests
werkzeug
rapidfuzz
numpy
faiss-cpu
sentence-transformers
google-generativeai
